+++ CHANGELOG.md
@@ -26,6 +26,10 @@ Security in case of vulnerabilities.
 - Add an "orderBy" argument to the `ListAllProjectsByFilter` query
 - Allow `ListAllProjectsByFilter` query to handle multiple filters
 
+### Changed
+ - Filter out any local authorities with no projects in `ListAllProjectByLocalAuthorities`
+
+ 
 See the [full commit history](https://github.com/DFE-Digital/complete-conversions-transfers-changes/compare/main...production-2025-04-17.164) for everything awaiting release
 
 ---
@@ -43,6 +47,10 @@ See the [full commit history](https://github.com/DFE-Digital/complete-conversion
  - Added missing "project for region" header
  - User redirection on app load based on their permissions
  - Add navigation items to be more consistent with ruby UI
+ - New route `/projects/all/by-month/conversions/{month}/{year}`
+ - New route `/projects/all/by-month/transfers/{month}/{year}`
+ - New route `/projects/all/by-month/conversions/from/{fromMonth}/{fromYear}/to/{toMonth}/{toYear}`
+ - New route `/projects/all/by-month/transfers/from/{fromMonth}/{fromYear}/to/{toMonth}/{toYear}`
 
 ### Changed
  - Updated route `/accessibility-statement` to `/accessibility`